added validation to delete method (now it returning 404 if there is no task with provided id in repository)

+++ src/main/java/com/app/YesItstodo/controller/TaskController.java
@@ -59,7 +59,10 @@ class TaskController {
     }
 
     @DeleteMapping("/tasks/{id}")
-    ResponseEntity<?> deleteTask(@PathVariable int id){
+    ResponseEntity<?> deleteTask(@PathVariable int id) {
+        if (!repository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         repository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
